refactor: re-design signup flow
