+++ src/pages/signup.tsx
@@ -14,52 +14,7 @@ import type { NextPageWithLayout } from "../pages/_app";
 import type { LoginFormType } from "@/types";
 
 const SignUpPage: NextPageWithLayout = () => {
-  const router = useRouter();
-  const dispatch = useDispatch();
-
-  const [statusCode, setStatusCode] = useState(0);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleSignUp = async (data: LoginFormType) => {
-    setIsLoading(true);
-    setStatusCode(0); // 重置狀態 否則 hook-form 的 error 會被清空
-
-    try {
-      const response = await fetch(apiNext.SIGN_UP, {
-        method: "POST",
-        body: JSON.stringify(data),
-      });
-
-      // 登入請求
-      if (response.ok) {
-        const response = await fetch(apiNext.LOGIN, {
-          method: "POST",
-          body: JSON.stringify(data),
-        });
-        const result = await response.json();
-
-        result.data && dispatch(setUserInfo(result.data));
-        router.push("/member/new/create_profile");
-      } else {
-        setStatusCode(response.status);
-      }
-    } catch (error) {
-      console.error("未知的錯誤");
-    }
-
-    setIsLoading(false);
-  };
-
-  return (
-    <>
-      <SignUp
-        isLoading={isLoading}
-        statusCode={statusCode}
-        onSubmit={handleSignUp}
-      />
-      {isLoading && <Loading />}
-    </>
-  );
+  return <SignUp />;
 };
 
 SignUpPage.getLayout = function getLayout(page: ReactElement) {
